There are a many clippy warnings in the main branch. This PR addresses
some of them.
I will fix the rest of the warnings in follow up PRs. Furthermore, the
clippy CI check is not working. I can look into that after I fixed the
rest of the clippy warnings. After that, I think it would make sense to
migrate to the latest Rust version. Does that work for you?
EDIT: This PR resolves all Clippy warnings and enables the CI to run all
Clippy checks. The CI has extra checks for the `limbo-wasm` module. This
module still has some Clippy warnings that need attention.


Closes #1742

+++ core/ext/vtab_xconnect.rs
@@ -61,7 +61,7 @@ pub unsafe extern "C" fn execute(
                         return ResultCode::Error;
                     }
                     Ok(StepResult::Done) => {
-                        *last_insert_rowid = conn.last_insert_rowid() as i64;
+                        *last_insert_rowid = conn.last_insert_rowid();
                         return ResultCode::OK;
                     }
                     Ok(StepResult::IO) => {
